// src/lib/types/cardano/network.ts
export type ProtocolParams = {
  a0: number,
  coinsPerUtxoWord: number,
  collateralPercent: number,
  costModel: number[],
  costModels: LanguageView,
  decentralisationParam: number,
  eMax: number,
  extraEntropy: string,
  keyDeposit: number,
  maxBlockBodySize: number,
  maxBlockExMem: string,
  maxBlockExSteps: string,
  maxBlockHeaderSize: number,
  maxCollateralInputs: number,
  maxTxExMem: string,
  maxTxExSteps: string,
  maxTxSize: number,
  maxValSize: string,
  minFeeA: number,
  minFeeB: number,
  minPoolCost: number,
  minUTxOValue: number,
  nOpt: number,
  poolDeposit: number,
  priceMem: number,
  priceStep: number,
  protocolVersion: {
    major: number,
    minor: number
  },
  rho: number,
  tau: number
}